Adjusted setup simulator to be used as a tool, whitout being attached to a Highline
